refactor: extract gravity_wells into analyzer/gravity_wells.rs (#190)
Moves GravityWell + compute_gravity_wells out of analyzer/mod.rs. apply_risk_weights() in mod.rs continues to call it through the re-export.
